feat(TodoContainer.js/TodoList.js/TodoListItem.js): add logic to move todo task up or down if checked.
This sets a state to manage wether a todo item is checked or not. It will push a newly checked item to the bottom of the list and it push a newly unchecked todo to the top of the list. Logic needs some work as the list gets jumbled when utilied along side the sort by title logic.